30. The while loop: more examples 3.2.3
Enter a number of type 0 to stop: 1
Enter a number of type 0 to stop: 2
Enter a number of type 0 to stop: 0
Odd numbers count:  1
Even number count:  1

Inside the loop. 5
Outside the loop. 4
Inside the loop. 4
Outside the loop. 3
Inside the loop. 3
Outside the loop. 2
Inside the loop. 2
Outside the loop. 1
Inside the loop. 1
Outside the loop. 0

Inside the loop. 5
Outside the loop. 4
Inside the loop. 4
Outside the loop. 3
Inside the loop. 3
Outside the loop. 2
Inside the loop. 2
Outside the loop. 1
Inside the loop. 1
Outside the loop. 0