34. break and continue statement 3.2.5
The break instruction:
---------------------- 
Inside the loop.  1
Inside the loop.  2
Outside the loop.

The continue instruction:
------------------------- 
Inside the loop:  1
Inside the loop:  2
Inside the loop:  4
Inside the loop:  5
Outside the loop.

Enter a number or type -1 to end the program: 98
Enter a number or type -1 to end the program: 78
Enter a number or type -1 to end the program: -1
The largest number is:  98

Enter a number or type -1 to end the program: 56
Enter a number or type -1 to end the program: 77
Enter a number or type -1 to end the program: -1
The largest number is:  77